v16.2.42 is released

+++ src/grid/actions/toolbar.ts
@@ -79,7 +79,7 @@ export class Toolbar {
      * @return {void}
      */
     public destroy(): void {
-        if (!this.toolbar.isDestroyed) {
+        if (this.toolbar && !this.toolbar.isDestroyed) {
             if (!this.toolbar.element) {
                 this.parent.destroyTemplate(['toolbarTemplate']);
             } else {
